Fix: valid if video is not nul

+++ apps/video/management/commands/clean_files.py
@@ -33,10 +33,11 @@ class Command(BaseCommand):
                     size_removed += os.path.getsize(url_file_proccessed)
                     os.remove(url_file_proccessed)
 
-                url_file_base = f'{str(settings.BASE_DIR)}{video.video.url}'
-                if os.path.isfile(url_file_base):
-                    size_removed += os.path.getsize(url_file_base)
-                    os.remove(url_file_base)
+                if video.video is not None:
+                    url_file_base = f'{str(settings.BASE_DIR)}{video.video.url}'
+                    if os.path.isfile(url_file_base):
+                        size_removed += os.path.getsize(url_file_base)
+                        os.remove(url_file_base)
 
                 video.video.delete()
                 video.save()
